+++ lib/blocs/habits/habits_cubit.dart
@@ -67,7 +67,7 @@ class HabitsCubit extends Cubit<HabitsState> {
   void startWatching() {
     _completionsStream = _journalDb.watchHabitCompletionsInRange(
       rangeStart: getStartOfDay(
-        DateTime.now().subtract(const Duration(days: 90)),
+        DateTime.now().subtract(const Duration(days: 180)),
       ),
     );