+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed:
 - Category save duplicate warning
 - Disable category save icon when form invalid
+- Recreate category with the same name as previously deleted category
 
 ### Changed:
 - Upgraded dependencies